+++ src/main/java/ru/belonogov/task_service/service/impl/TaskServiceImpl.java
@@ -8,7 +8,10 @@ import ru.belonogov.task_service.domain.dto.response.TaskResponse;
 import ru.belonogov.task_service.domain.entity.Task;
 import ru.belonogov.task_service.domain.entity.TaskStatus;
 import ru.belonogov.task_service.domain.exception.AddNewEmployeeException;
+import ru.belonogov.task_service.domain.exception.AddNewTaskException;
 import ru.belonogov.task_service.domain.exception.TaskNotFoundException;
+import ru.belonogov.task_service.domain.exception.UpdateException;
+import ru.belonogov.task_service.domain.repository.EmployeeDao;
 import ru.belonogov.task_service.domain.repository.TaskDao;
 import ru.belonogov.task_service.service.TaskService;
 
@@ -20,10 +23,12 @@ public class TaskServiceImpl implements TaskService {
 
     private final TaskMapper taskMapper;
     private final TaskDao taskDao;
+    private final EmployeeDao employeeDao;
 
-    public TaskServiceImpl(TaskMapper taskMapper, TaskDao taskDao) {
+    public TaskServiceImpl(TaskMapper taskMapper, TaskDao taskDao, EmployeeDao employeeDao) {
         this.taskMapper = taskMapper;
         this.taskDao = taskDao;
+        this.employeeDao = employeeDao;
     }
 
     @Override
@@ -50,6 +55,9 @@ public class TaskServiceImpl implements TaskService {
     @Override
     public List<TaskResponse> findAllByEmployee(Long id) {
         List<Task> allByEmployee = taskDao.findAllByEmployee(id);
+        if(allByEmployee.isEmpty()) {
+            return Collections.emptyList();
+        }
         List<TaskResponse> result = allByEmployee.stream()
                 .map($ -> taskMapper.taskToTaskResponse($))
                 .collect(Collectors.toList());
@@ -65,6 +73,9 @@ public class TaskServiceImpl implements TaskService {
         task.setDescription(taskUpdateRequest.getDescription());
         task.setRating(taskUpdateRequest.getRating());
         task.setTaskStatus(taskUpdateRequest.getTaskStatus());
+        if(taskDao.findById(task.getId()).isEmpty()) {
+            throw new UpdateException("Задание не найдено");
+        }
         Task update = taskDao.update(task);
 
         return taskMapper.taskToTaskResponse(update);
@@ -74,13 +85,17 @@ public class TaskServiceImpl implements TaskService {
     public void addNewEmployeeToTask(TaskEmployeeRequest taskEmployeeRequest) {
         Long taskId = taskEmployeeRequest.getTaskId();
         Long employeeId = taskEmployeeRequest.getEmployeeId();
-        if(!taskDao.addNewEmployeeToTask(taskId, employeeId)) {
-            throw new AddNewEmployeeException("Работник не был назначен на задание");
+        if(taskDao.findById(taskId).isEmpty() || employeeDao.findById(employeeId).isEmpty()) {
+            throw new AddNewTaskException("Ошибка добавления нового задания работнику");
         }
+        taskDao.addNewEmployeeToTask(taskId, employeeId);
     }
 
     @Override
-    public void delete(Long id) {
-        taskDao.delete(id);
+    public boolean delete(Long id) {
+        if (taskDao.findById(id).isEmpty()){
+            throw new UpdateException("Задание не найдено");
+        }
+        return taskDao.delete(id);
     }
 }
